Added additional hashmap methods: has, remove
has(key): takes a key as an argument and returns true or false based on
whether or not the key is in the hash map.

remove(key): takes a key as an argument. If the given key is in the hash
 map, it should remove the entry with that key and return true. If the
 key isn't in the hash map, it should return false.

removeAt(key): this was updated to handle linkedList of length = 1

 Changes to be committed:
	modified:   hashMap.js
	new file:   index.html