Added hashmap methods: length, clear
length(): the data structure was altered to store a counter variable
which updates whenever nodes are added, removed, or if the hashmap is
cleared.

clear(): First, within the linkedlist a clear function was added which
makes both head and tail equal null. Then within the hashmap, a clear
function was added which iterates throughout the and call the linkedlist
method to clear.

 Changes to be committed:
	modified:   hashMap.js